# Changes toward `layernorm_forward` in `dev/cuda`