@karpathy 2024-06-16
Hi, this looks ok to me.

- the coding style in kernel6 is a bit off, e.g. i think you're using black? the cropping of the lines and such looks inconsistent with the rest of the code.
- am i reading this right that kernel6 is fairly fast? would it make sense to incorporate it into prod?
- cc @ngc92 or @ademeure to also possibly take a look