@KarhouTam 2024-06-16
Hi, thanks for reviewing this PR.

> the coding style in kernel6 is a bit off, e.g. i think you're using black? the cropping of the lines and such looks inconsistent with the rest of the code.

I used `clang-format` and the .clang-format is based on google code style. It's ok to re-format the codes (by myself or others). But I don't have llm.c's format rule file.

> would it make sense to incorporate it into prod?

My machine doesn't have that much computaional resource to do the training. So please do this test for me. If it cannot be used in production, rolling back commits is acceptable.