@KarhouTam 2024-06-16
> the coding style in kernel6 is a bit off, e.g. i think you're using black? the cropping of the lines and such looks inconsistent with the rest of the code.

I have tried to restore the code style of former kernels. Does it look better now?