@KarhouTam 2024-06-17
Hi, @karpathy. 

I've tried to run `profile_gpt2.cu` to profile currently used layernorm_forward kernel (`layernorm_forward_kernel3`) in `layernorm.cuh` and my kernel (`layernorm_forward_kernel4`).

### 512 block size (current)
[full-ncu-results](https://github.com/karpathy/llm.c/assets/82705454/612ed424-a39c-48c8-a705-34f02f3cc831)

My kernel performs not that good when block size is 512. Since my gpu is 3070 laptop (CC 8.6) and smem is not enough, thus the bottleneck may be smem capacity. 

Brief results:
![image](https://github.com/karpathy/llm.c/assets/82705454/6e2eb56a-95d3-43b4-af8a-e99a2934214b)

### 256 block size
[full-ncu-results](https://github.com/karpathy/llm.c/assets/82705454/075fdcd1-32d8-4862-b525-586da9062322)
But my kernel and the current performs better with block size 256. So maybe the block size set in `layernorm_forward()` is not the best option.

Brief results:
![image](https://github.com/karpathy/llm.c/assets/82705454/2f49eb1e-99f1-467d-a119-84bead3ce04b)