@KarhouTam 2024-06-18
Notice that the name kernel6 is taken, thus rename the kernel to kernel7.

Meanwhile combines the merits of kernel4 and kernel6. kernel7 achieves similar performance to kernel6 and kernel4 wihle using less shared memory.

### benchmark

#### `layernorm_forward_kernel6`
```
block_size   32 | time 0.1288 ms | bandwidth 390.77 GB/s
block_size   64 | time 0.1282 ms | bandwidth 392.56 GB/s
block_size  128 | time 0.1284 ms | bandwidth 391.89 GB/s
block_size  256 | time 0.1275 ms | bandwidth 394.67 GB/s
block_size  512 | time 0.1282 ms | bandwidth 392.51 GB/s
block_size 1024 | time 0.3942 ms | bandwidth 127.68 GB/s
```
P.S. the poor performance of block size 1024 maybe due to my poor 3070 laptop gpu.

#### `layernorm_forward_kernel7`
```
block_size   32 | time 0.1285 ms | bandwidth 391.67 GB/s
block_size   64 | time 0.1271 ms | bandwidth 396.08 GB/s
block_size  128 | time 0.1271 ms | bandwidth 395.85 GB/s
block_size  256 | time 0.1271 ms | bandwidth 396.12 GB/s
block_size  512 | time 0.1281 ms | bandwidth 392.83 GB/s
block_size 1024 | time 0.1283 ms | bandwidth 392.16 GB/s
```