@karpathy 2024-06-18
Hi @KarhouTam we just merged a LayerNorm forward, I'm not 100% sure how this version is similar or different now.