@KarhouTam 2024-06-18
Hi, @karpathy. Notice that kernel6 use smem to store data from input, weight and bias. In the process of layernorm forward, the weight and bias data are only used at the end (calculate the final result). So I think that is unnecessary to store them by smem the beginning.